requirements.txt
flask==2.0.2
marshmallow==3.14.1
sqlalchemy==1.4.29
#alembic==1.7.5 
mysqlclient==2.1.0
flask_sqlalchemy==2.5.1
flask_migrate==3.1.0
python-dotenv==0.19.2